requirements.txt
# Data Scraping
google-api-python-client>=2.100.0
praw>=7.7.0
feedparser>=6.0.0
pytrends>=4.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# NLP & Embeddings
openai>=1.12.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Clustering & Math
scikit-learn>=1.3.0
hdbscan>=0.8.33
numpy>=1.24.0
pandas>=2.1.0

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.6.0

# Supabase / PostgreSQL
supabase>=2.3.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# GCP
google-cloud-tasks>=2.14.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0
tqdm>=4.66.0
